Wip v2.01 (#1)
* Rework readme

* Edit readme content

* Clarified and simplified README

* Remove bizaare Hash structure of cart

* Add specs for find_item_by_name_in_collection

This is a helper method that helps with the
location of a Hash within an Array. Students don't
yet know Enumerables or Hash-pair iteration, so
they really need this "boost."

* Remove Enumerable use in specs

Inappropriate for our laziness and (to be fair)
DRY-ness to needlessly confuse the matter.

* Consequence of de-Hashing the carts

* De-DRY before-blocks

Confusing for learners at this stage. But the
WETness does make me sad.

* Formatting

* Remove purchase-penalizing test

Seems real strange. If someone buys > coupon count
of items, cool, we should just charge them coupon
price and regular price. That seems the whole
point of keeping track of count in the
consolidated cart. If you don't meet the number,
yes, you should not get the coupon's power. But
more...seems wrong.

Succinctly: What grocer is going to complain that
people bough _more_?

* Remove inappropriate spec fanciness

* Remove test where multiple coupons come round

We've exercised them enough! :)

It's also hard to do without Enumerables so let's
cut that complexity.

* Clean up use of apply_clearance

It was unclear whether apply_clearance was
mutating or not. In the README we now say,
"mutate," so adjust specs to not require a return
value for success.

Also remove Enumerable fanciness.

Also, use a better matcher for testing Floats.

* Remove fanciness

* Remove to receive calls

This is so flakey and confusing, I think we're
better off, for a beginner audience not working in
a `class` context simply documenting the
expectation.

We set the expectation in both comments and the
README.

* Remove Hash-based thinking

* Remove stray receiver test

* Remove illegit. helper find_coupon

* Clarify retval non-important

* rm unused vars

* Weird case test

* Formatting

* Remove illegit find_item helper

* rm confusing Array with defalu value invocations

* Baseline grocer impl file

* Remove Hash-structure based calls

* Clarify learning goals

* fixes some grammar, clarifies checkout instruction

* fix code snippet errors

* Grammar

* Add expec: find_item_by_name_in_collection

Test the positive case

* Wip no inplace methods (#2)

* Standardize Spec: apply_coupon()'s variable

Make things easier by having apply_coupon() always
return coupon_applied_cart().

* Remove unused call to apply_coupons

* Standardize spec: consolidated_cart

Make tests easier to read for beginners. Name
consolidate_cart()'s retvalue consistently.

* Standardize calls to apply_clearance

* Simplify checkout methods

* w/ coupon -> W/COUPON to match tests